Refactor DefendantAccountApiController and Service to streamline history retrieval with direct parameters
